% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{boot.BCa}
\alias{boot.BCa}
\title{BCa bootstrap confidence interval.}
\usage{
boot.BCa(x, th0, th, stat, conf = 0.95)
}
\arguments{
\item{x}{the observations}

\item{th0}{the observed statistic.}

\item{th}{is the vector of bootstrap replicates.}

\item{stat}{the function to compute the statistic}

\item{conf}{the confidence level}
}
\value{
the maximum number of extreme points
}
\description{
BCa bootstrap confidence interval.
}
