% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cal.quant}
\alias{cal.quant}
\title{quantile.}
\usage{
cal.quant(x)
}
\arguments{
\item{x}{the sample}
}
\value{
max, min, median, upper and lower quartile
}
\description{
calculate max, min, median, upper and lower quartile.
}
\examples{
\dontrun{
set.seed(123)
x <- rnorm(100, 1, 2)
my.fivenum(x)
}
}
