% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{calcCI}
\alias{calcCI}
\title{calculate confidence interval.}
\usage{
calcCI(n, alpha)
}
\arguments{
\item{n}{the number of samples.}

\item{alpha}{significance level.}
}
\value{
confidence border
}
\description{
A function used to calculate confidence interval.
}
\examples{
\dontrun{
C1 <- replicate(1000, expr = calcCI(n = 20, alpha = .025))
C2 <- replicate(1000, expr = calcCI(n = 20, alpha = .975))
print(c(exp(mean(C1)), exp(mean(C2)))) ## the confidence interval
}
}
