% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chisq.var.test}
\alias{chisq.var.test}
\title{chisquqre-estimating interval of population variance.}
\usage{
chisq.var.test(x, alpha)
}
\arguments{
\item{x}{the sample}

\item{alpha}{the significance level}
}
\value{
interval estimation of population variance
}
\description{
estimating interval of population variance using chi-square statistics.
}
\examples{
\dontrun{
set.seed(123)
x <- rnorm(100, 1, 2)
chisq.var.test(x, 0.05)
}
}
