% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{count5test}
\alias{count5test}
\title{count five method for hypothesis tests.}
\usage{
count5test(x, y)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a vector, matrix or data frame with compatible dimensions to x.}
}
\value{
the value 1 (reject null hypothesis) or 0 (do not reject null hypothesis)
}
\description{
count five method for hypothesis tests.
}
\examples{
\dontrun{
power <- mean(replicate(1e3, expr={
  x <- rnorm(20, 0, 1)
  y <- rnorm(20, 0, 1.5)
  count5test(x, y)
}))
}
}
