% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{the Gelman-Rubin method of monitoring convergence of a Metropolis chain.}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{the proposal distribution.}
}
\value{
The Gelman-Rubin statistic
}
\description{
the Gelman-Rubin method of monitoring convergence of a Metropolis chain.
}
