% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{Mardia.test}
\alias{Mardia.test}
\title{Mardia tests of multivariate normality.}
\usage{
Mardia.test(X)
}
\arguments{
\item{X}{the observation.}
}
\value{
the empirical estimates of power
}
\description{
tests of multivariate normality based on multivariate generalizations of skewness and kurtosis proposed by Mardia.
}
