% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{maxout}
\alias{maxout}
\title{the maximum number of extreme points.}
\usage{
maxout(x, y)
}
\arguments{
\item{x}{numeric or complex or logical vectors.}

\item{y}{numeric or complex or logical vectors.}
}
\value{
the maximum number of extreme points
}
\description{
counts the maximum number of extreme points of each sample with respect to the range of the other sample.
}
\examples{
\dontrun{
x <- rnorm(10, 0, 1)
y <- rnorm(10, 0, 1)
res <- maxout(x, y)
}
}
