% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MetropolisC}
\alias{MetropolisC}
\title{A Metropolis sampler using Rcpp}
\usage{
MetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{vector of standard deviations.}

\item{x0}{initial value.}

\item{N}{the length of the chain.}
}
\value{
a random sample of size
}
\description{
A Metropolis sampler using Rcpp
}
