% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{MetropolisR}
\alias{MetropolisR}
\title{A Metropolis sampler using R.}
\usage{
MetropolisR(n, sigma, x0, N)
}
\arguments{
\item{n}{the number of the samples.}

\item{sigma}{vector of standard deviations.}

\item{x0}{initial value.}

\item{N}{the length of the chain.}
}
\value{
a random walk Metropolis sample of size
}
\description{
A Metropolis sampler using R.
}
\examples{
\dontrun{
rw <- MetropolisR(4, 2, 25, 2000)
plot(rw$x, type = "l")
}
}
