% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{nlapply}
\alias{nlapply}
\title{calculate confidence interval.}
\usage{
nlapply(x1, x2, FUN, FUN.VALUE, USE.NAMES = TRUE)
}
\arguments{
\item{x1}{A list or atomic vector.}

\item{x2}{A list or atomic vector.}

\item{FUN}{the function to be applied to each element of X.}

\item{FUN.VALUE}{a (generalized) vector; a template for the return value from FUN. See ‘Details’.}

\item{USE.NAMES}{logical.}
}
\description{
A function used to calculate confidence interval.
}
\examples{
\dontrun{
set.seed(123)
xs <- list(runif(10), runif(20))
ws <- list(rpois(10, 2) + 1, rpois(20, 2) + 1)
res <- nlapply(xs,ws,weighted.mean, numeric(1))
}
}
