% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{normal.chain}
\alias{normal.chain}
\title{The Metropolis-Hastings sampler.}
\usage{
normal.chain(sigma, N, X1)
}
\arguments{
\item{sigma}{standard deviation of proposal distribution}

\item{N}{length of the chain}

\item{X1}{mean of proposal distribution}
}
\value{
confidence border
}
\description{
The Metropolis-Hastings sampler.
}
