% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{pwr_t}
\alias{pwr_t}
\title{the empirical power using  \code{t(v)}}
\usage{
pwr_t(v)
}
\arguments{
\item{v}{degrees of freedom}
}
\value{
the empirical power
}
\description{
calculate the empirical power with heavy-tailed t.
}
