% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RayleighC}
\alias{RayleighC}
\title{A Rayleigh sampler using Rcpp}
\usage{
RayleighC(n, sigma)
}
\arguments{
\item{n}{size of Rayleigh random number}

\item{sigma}{parameter of rayleigh density}
}
\value{
a Rayleigh random sample
}
\description{
A Rayleigh sampler using Rcpp
}
