% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RayleighR}
\alias{RayleighR}
\title{Rayleigh sampler}
\usage{
RayleighR(n, sigma)
}
\arguments{
\item{n}{size of the random number}

\item{sigma}{parameter of rayleigh density}
}
\value{
Rayleigh random number and iteration times
}
\description{
Rayleigh sampler using acceptance-rejection method
}
\examples{
\dontrun{
n <- 1000
sigma <- 1
x <- Rayleigh(n, sigma)
}
}
