% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{sk}
\alias{sk}
\title{skewness.}
\usage{
sk(x)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}
}
\value{
the skewness of the sample
}
\description{
skewness.
}
\examples{
\dontrun{
x <- rnorm(100)
res <- sk(x)
}
}
