% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{theta.boot}
\alias{theta.boot}
\title{function to compute the statistic.}
\usage{
theta.boot(dat, ind)
}
\arguments{
\item{dat}{the observed data}

\item{ind}{the observed data}
}
\value{
a simulated data set of the same form as the observed data
}
\description{
function to compute the statistic.
}
